Adicionado Filtro por preço para AppointmentType

+++ Backend/fila-flex/src/main/java/com/qmasters/fila_flex/service/AppointmentTypeService.java
@@ -11,6 +11,8 @@ import com.qmasters.fila_flex.dto.CategoryDTO;
 import com.qmasters.fila_flex.model.AppointmentType;
 import com.qmasters.fila_flex.repository.AppointmentTypeRepository;
 
+import jakarta.transaction.Transactional;
+
 @Service
 public class AppointmentTypeService {
 
@@ -40,7 +42,7 @@ public class AppointmentTypeService {
             categoryService.saveCategory(categoryDTO);
         }
 
-        return toDTO(appointmentType);
+        return toDTO(appointmentType); //talvez mudar isso para retornar o model ao invés do DTO
     }
 
     public List<AppointmentType> listAll() {
@@ -51,55 +53,59 @@ public class AppointmentTypeService {
         return appointmentTypeRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("AppointmentType não encontrado."));
     }
-/* 
-    public AppointmentTypeDTO findById(Long id) {
-        return appointmentTypeRepository.findById(id)
-                .map(this::toDTO)
-                .orElseThrow(() -> new RuntimeException("AppointmentType não encontrado."));
-    }*/
-
-    //Método para buscar AppointmentTypes filtrados por categoria.
-    public List<AppointmentTypeDTO> findByCategory(String category) {
-        return appointmentTypeRepository.findByCategory(category).stream()
-                .map(this::toDTO)
-                .collect(Collectors.toList());
-    }
-
-    //Método para buscar AppointmentTypes por intervalo de preços.
-    public List<AppointmentTypeDTO> findByPriceBetween(double minPrice, double maxPrice) {
-        return appointmentTypeRepository.findByPriceBetween(minPrice, maxPrice).stream()
-                .map(this::toDTO)
-                .collect(Collectors.toList());
-    }
-
-    public void deleteAppointmentType(Long id) {
-        appointmentTypeRepository.deleteById(id);
-    }
-
-    //Método para deletar AppointmentType por nome.
-    public void deleteByName(String name) {
-        // Encontra o AppointmentType pelo nome
-        AppointmentType appointmentType = appointmentTypeRepository.findByName(name)
-            .orElseThrow(() -> new RuntimeException("Tipo de agendamento não encontrado com o nome: " + name));
-        
-        // Deleta o AppointmentType
-        appointmentTypeRepository.delete(appointmentType);
-    }
-
+    
     public AppointmentType findByName(String name) {
         return appointmentTypeRepository.findByName(name)
                 .orElseThrow(() -> new RuntimeException("AppointmentType não encontrado."));
     }
 
+    // buscar AppointmentTypes filtrados por categoria.
+    public List<AppointmentTypeDTO> findByCategory(String category) {
+        return appointmentTypeRepository.findByCategory(category).stream()
+        .map(this::toDTO)
+        .collect(Collectors.toList());
+    }
+
+    //buscar AppointmentTypes por intervalo de preços.
+    public List<AppointmentTypeDTO> findByPriceBetween(double minPrice, double maxPrice) {
+        return appointmentTypeRepository.findByPriceBetween(minPrice, maxPrice).stream()
+                .map(this::toDTO)
+                .collect(Collectors.toList());
+    }
+    
+    //buscar AppointmentTypes por ordem de tempo estimado.
+    public List<AppointmentType> findAllByOrderByEstimatedTimeAsc() {
+        return appointmentTypeRepository.findAllByOrderByEstimatedTimeAsc();
+    }
+    
+    public void deleteByName(String name) {
+        AppointmentType appointmentType = appointmentTypeRepository.findByName(name)
+        .orElseThrow(() -> new RuntimeException("Tipo de agendamento não encontrado com o nome: " + name));
+        
+        appointmentTypeRepository.delete(appointmentType);
+    }
+    
+
+    @Transactional
+    public void deleteAppointmentType(Long id) {
+        if (appointmentTypeRepository.existsById(id)) {
+            appointmentTypeRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("AppointmentType não encontrado.");
+        }
+    }
+    
     private AppointmentTypeDTO toDTO(AppointmentType appointmentType) {
-        return new AppointmentTypeDTO(appointmentType.getName(),
-                appointmentType.getDescription(),
-                appointmentType.getCategory(),
-                appointmentType.getPrice(),
-                appointmentType.getEstimatedTime(),
-                appointmentType.getAppointmentDate(),
-                appointmentType.getRequiredDocumentation(),
-                appointmentType.getAdress()
+        return new AppointmentTypeDTO(
+            appointmentType.getName(),
+            appointmentType.getDescription(),
+            appointmentType.getCategory(),
+            appointmentType.getPrice(),
+            appointmentType.getEstimatedTime(),
+            appointmentType.getAppointmentDate(),
+            appointmentType.getRequiredDocumentation(),
+            appointmentType.getAdress()
         );
     }
+
 }
